test: split breadcrumb tests into jvm vs serialisation tests
